refactor(core): extract Opus packet framing
Move RFC 6716 TOC / code 0–3 frame slicing into
audio/opus/framing.rs. The decoder still rewrites each frame as
a code 0 packet; public decode paths are unchanged.
